+++ app/Support/Http/Controllers/UserNavigation.php
@@ -97,8 +97,8 @@ trait UserNavigation
         /** @var Transaction $transaction */
         foreach ($transactions as $transaction) {
             $type = $transaction->account->accountType->type;
-            if (!in_array($type, $ignore)) {
-                return redirect(route('accounts.show', [$transaction->account_id]));
+            if (!in_array($type, $ignore, true)) {
+                return redirect(route('accounts.edit', [$transaction->account_id]));
             }
         }