+++ resources/lang/es_ES/breadcrumbs.php
@@ -51,6 +51,7 @@ return [
     'edit_journal'           => 'Editar transacción ":description"',
     'edit_reconciliation'    => 'Editar ":description"',
     'delete_journal'         => 'Eliminar transacción ":description"',
+    'delete_group'           => 'Delete transaction ":description"',
     'tags'                   => 'Etiquetas',
     'createTag'              => 'Crear nueva etiqueta',
     'edit_tag'               => 'Editar etiqueta ":tag"',