+++ resources/lang/it_IT/import.php
@@ -38,6 +38,7 @@ return [
 
     // notices about the CSV importer:
     'deprecate_csv_import' => 'Come descritto in <a href="https://www.patreon.com/posts/future-updates-30012174">questo post Patreon</a>, il modo in cui Firefly III gestisce l\'importazione dei dati cambierà. Ciò significa che l\'importatore CSV sarà spostato in un nuovo strumento separato. Puoi già testare questo strumento visitando <a href="https://github.com/firefly-iii/csv-importer">questo repository GitHub</a>. Prova il nuovo importatore e fammi sapere cosa ne pensi.',
+    'final_csv_import'     => 'Come descritto in <a href="https://www.patreon.com/posts/future-updates-30012174">questo post Patreon</a>, il modo in cui Firefly III gestisce l\'importazione dei dati cambierà. Ciò significa che questa è l\'ultima versione di Firefly III che presenta l\'importatore CSV. È disponibile uno strumento separato che puoi provare tu stesso: <a href="https://github.com/firefly-iii/csv-importer">l\'importatore CSV di Firefly III</a>. Ti sarei molto grato se lo provassi e mi facessi sapere cosa ne pensi.',
 
     // import provider strings (index):
     'button_fake'                         => 'Esegui un\'importazione fittizia',