+++ resources/lang/nl_NL/import.php
@@ -38,6 +38,7 @@ return [
 
     // notices about the CSV importer:
     'deprecate_csv_import' => 'De manier waarop Firefly III je data laat importeren gaat veranderen. Je kan lezen in <a href="https://www.patreon.com/posts/future-updates-30012174">deze Patreon-post</a> dat de CSV import-tool gaat verhuizen naar een zelfstandige repository en tool. Je kan deze alvast beta-testen als je <a href="https://github.com/firefly-iii/csv-importer">deze GitHub repository</a> bezoekt. Als je dat zou willen doen, heel graag, en laat me weten of het lukt.',
+    'final_csv_import'     => 'De manier waarop Firefly III je data laat importeren gaat veranderen. Je kan lezen in <a href="https://www.patreon.com/posts/future-updates-30012174">deze Patreon-post</a> dat de CSV import-tool gaat verhuizen naar een zelfstandige repository en tool. Dit is de laatste versie met de CSV importer. Check <a href="https://github.com/firefly-iii/csv-importer">deze GitHub repository</a> en test de nieuwe tool.',
 
     // import provider strings (index):
     'button_fake'                         => 'Nepdata importeren',