+++ resources/lang/pt_BR/breadcrumbs.php
@@ -51,6 +51,7 @@ return [
     'edit_journal'           => 'Editar transação ":description"',
     'edit_reconciliation'    => 'Editar ":description"',
     'delete_journal'         => 'Apagar transação ":description"',
+    'delete_group'           => 'Delete transaction ":description"',
     'tags'                   => 'Etiquetas',
     'createTag'              => 'Criar nova etiqueta',
     'edit_tag'               => 'Editar etiqueta ":tag"',